moved interface to models folder - add h2 page title styling
