+++ src/app/core/models/campaign.inteface.ts
@@ -1,0 +1,8 @@
+export interface Campaign {
+  title: string;
+  description: string;
+  date: Date;
+  place: string;
+  img: string;
+  imgTitle: string;
+}
