+++ src/app/features/public/pages/campaigns/campaigns.component.scss
@@ -7,6 +7,25 @@
   }
 }
 
+.campaign-title > h2 {
+  text-transform: uppercase;
+  letter-spacing: 4px;
+  font-size: 2.2em;
+  text-shadow: 1px 2px 2px rgba(95, 95, 95, 0.89);
+  color: $gold;
+
+  @media screen and (max-width: 991px) {
+    //medium devices
+    font-size: 1.5em;
+  }
+
+  @media screen and (max-width: 575px) {
+    // small devices
+    letter-spacing: 2px;
+    font-size: 1em;
+  }
+}
+
 .image-container {
   max-height: 60vh;
   & img {
